requirements.txt
Django==3.0.7
# django-bootstrap4
psycopg2
psycopg2-binary
uwSGI==2.0.18
djangorestframework==3.11.0
django-cors-headers==3.4.0
# mysqlclient==1.4.6